+++ http4k-contract/src/test/kotlin/org/http4k/contract/OpenApiTest.kt
@@ -1,0 +1,5 @@
+package org.http4k.contract
+
+import org.http4k.format.Argo
+
+class OpenApiTest : ContractRendererContract(OpenApi(ApiInfo("title", "1.2", "module description"), Argo))